Use AbbreviatedKey in mergingIter
This commit adds support for AbbreviatedKey to the mergingIter.

It also corrects the comment above the definition of AbbreviatedKey,
changing "iff" to implication, since a < b does not imply
AbbreviatedKey(a) < AbbreviatedKey(b) (AbbreviatedKey is monotonic).

It also fixes a bug in the definition of AbbreviatedKey for the
DefaultComparer, which wouldn't align the abbreviated keys.